app/o_auth/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
redis
httpx
python-jose[cryptography]
python-dotenv